% Generated by build script
\name{awsip_ipv6_ranges}
\alias{awsip_ipv6_ranges}
\title{Get AWS IPv6 IP ranges}
\usage{
awsip_ipv6_ranges()
}
\arguments{
\item{service}{Filter by AWS service. Default NULL returns all.}
\item{region}{Filter by region pattern. Default NULL returns all.}
}
\value{tibble: ipv6_prefix, region, service, network_border_group}
\description{Get AWS IPv6 IP ranges}
