% Generated by build script
\name{awsip_ranges}
\alias{awsip_ranges}
\title{Get AWS IPv4 IP ranges}
\usage{
awsip_ranges()
}
\arguments{
\item{service}{Filter by AWS service (e.g. "S3", "EC2", "CLOUDFRONT",}
\item{region}{Filter by region pattern (e.g. "us-east-1", "eu-").}
}
\value{tibble: ip_prefix, region, service, network_border_group}
\description{Get AWS IPv4 IP ranges}
